Move trajectory endpoint to frame-trajectory endpoint
