+++ backend/__main__.py
@@ -3,13 +3,15 @@ import logging
 from flask import Flask, render_template
 from flask_cors import CORS
 
-from app.trajectory.controller import trajectory_controller
+from app.frame_trajectory.controller import frame_trajectory_controller
 
-app = Flask(__name__, static_folder="../frontend/dist/", template_folder="../frontend/dist/")
+app = Flask(__name__, static_folder="../frontend/dist/",
+            template_folder="../frontend/dist/")
 cors = CORS(app, resources={"/": {"origins": "*"}})
 app.config['CORS_HEADERS'] = 'Content-Type'
 
-app.register_blueprint(trajectory_controller, url_prefix='/trajectory')
+app.register_blueprint(frame_trajectory_controller,
+                       url_prefix='/frame-trajectory')
 
 
 @app.route('/', defaults={'path': ''})
